refactor(frontend): 拆分 AppLayout 上帝组件，统一组件化架构
- AppLayout 从 837 行精简至 228 行，仅保留导航/布局/全局状态
- 提取 6 个 workspace 组件（Background/Loading/UploadStage/ReviewStage/AnswerInputModal/FilterPanel）
- 创建独立 WorkspaceView 管理录入工作台全流程
- 8 个子视图自包含 ContentPanel，消除 props 透传
- 4 个 composable 改为单例模式（SystemStatus/ChatSession/ImageModal/WorkspaceNav）
- 新增 3 个复用 composable（useSelectableList/usePaginatedList/useToast）
- 目录重命名：views/workspace → views/app，Layout 文件归入对应目录
