+++ frontend/src/views/app/SplitHistoryView.vue
